+++ spec/integrations/pro/consumption/virtual_partitions/many_batches.rb
@@ -3,6 +3,7 @@
 # When using virtual partitions, we should easily consume data with the same instances on many
 # batches and until there is a rebalance or critical error, the consumer instances should
 # not change
+
 setup_karafka do |config|
   config.license.token = pro_license_token
   config.concurrency = 10